llm_backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
openai
google-generativeai
